Remove the retired AkoNotes and Notes resources
With the deprecation of KJots and KNotes applications, it's no longer
possible to create or access notes stored in Akonadi. For that reason,
we can retire and remove the now unused notes-related resources.

As a migration path for users, the Marknote application has gained
support to import notes from KJots/KNotes storage, and it's now the
recommended KDE application to use for note-taking.

See pim-technical-roadmap#58.