Forward network error to the resouce
job->error and job->errorString are always empty and so the error was
not forwarded. Due to it not being forwarded, this made the EWS request
believe it was successful, which then crashed due to the response data
being empty.

BUG: 485799
BUG: 484579
Fixed-In: 24.08.0


(cherry picked from commit 586bbacf6990e5d8264a4a0c5884f03458a8d969)

Co-authored-by: Carl Schwan <carl@carlschwan.eu>