Attempt to workaround EWS servers sometimes returning broken iCal
Apparently it's a known issue that some versions of EWS (including Outlook365)
in some situations return an invalid iCal that is missing the closing
END:VCALENDAR entry. This causes the event to be missing at best or to
crash the EWS code at worst.

This workaround simply checks for presence of END:VCALENDAR in the response
and if it's missing it will simply append it to the end.

BUG: 491681
CCBUG: 492120
FIXED-IN: 24.08.2