EWS: don't do autodiscovery when using Office365 OAuth
It will just work. Only perform autodiscovery when using username/
password auth or when user customizes the base url for OAuth2.

The autodiscovery on the default office365 address fails
for some reason, which I need to investigate further, howerver this
at least fixes the user exerience for majority of users who are
connecting to the Office365 instance.