Dav: fix updating categories on tags change
When a tag is added or removed, we also need to modify the payload of the
Item, since the tags are stored as event categories. After this, the changed
Item must be written back to Akonadi, otherwise Akonadi would still be showing
the old categories (at least until next sync).