agents/newmailnotifier: use <br/> for notifications
The notification specification, strictly speaking, does not support
newline tags at the moment, although some implementations like Plasma's
do support them. The specification does say that it is an xml-based
format though, and the newline tag that newmailnotifier currently sends,
<br> is only valid html, but not valid xml. This should not cause real
issues, as clients are supposed to ignore them anyway, but may cause
xml-based parsers to complain, and following the intent of the spec
seems better.

This change replaces <br> with <br/> in text sent to the notification
server.