urlconfigurationdialog.ui - move the protocol buttons into a QButtonGroup
Put the radio buttons for CalDav, CardDav and GroupDav into
a QButtonGroup, thereby always setting one DAV protocol.

Previously, it was possible (and by default) to have no
protocols selected when the user pressed Ok.  And that
caused a crash in DavProtocolBase *DavManager::davProtocol
when indexing the d->mProtocols[protocol] array.

BUG: 455787