Add Microsoft InTune authentication to EWS resource
This adds a new authentication option that uses Microsoft InTune for
Linux. This requires that the InTune application is provisioned and
the machine is joined to the organization's Azure directory.

The InTune authentication interacts with the Microsoft Identity
Broker, which is the authentication agent used by InTune. This broker
holds the device credentials and acts as a proxy for requesting OAuth2
access tokens.

Both interactive and non-interactive token requests are supported. The
interactive option uses a browser built into the identity broker and
therefore use of local web browser window is not necessary.

Currently this implementation is restricted to Unix-like platforms as
it doesn't make sense on Windows, which has its own native APIs for
this.

Signed-off-by: Krzysztof Nowicki <krissn@op.pl>