singlefileresourceconfigwidget.h - saving a empty path is a failure condition
Attempting to save an empty path is a failure condition.
This can happen in the ical (single file) resource when the Ok/Apply
buttons are pressed and the urlrequester is empty.

(Ideally, the Ok and Apply buttons would be disabled in the
empty urlrequester case, but the logic for that is entwined
in many different scenarios and resources.)

With this patch, if the user presses Ok/Apply with empty url
the dialog pops-down harmlessly.