WIP - assign an account icon to some well known providers
This is so we can show icons on the calendars and calendar items
to more easily distinguish which resource they belong to.

We already have this for the google resource, but all the other
resources have the same icon "folder-remote".

Add the icons also to the setupwizard combobox.

Can we propagate the setupwizard icons into the configuration dialog?

It will be quite some work to track down icons for all the providers.