Assign an account icon to some well known providers
This is so we can show icons on the calendars and calendar items
to more easily distinguish which resource they belong to.

We already have this for the google resource, but all the other
resources have the same icon "folder-remote".

Adds the icons to the service desktop files and uses them
in the setupwizard combobox; which are then propogated up into
the configuration dialog. If there isn't an icon found or if
manual configuration was used => use heuristics to determine one.

It will be quite some work to track down icons for all the providers.