IMAP: avoid doing SEARCH when we can just FETCH by sequence numbers
The SEARCH during sync was introduced as an optimization for the batching.
Its purpose was to get list valid UIDs within a sequence so that each
batch of UIDs is fully utilized. The problem with SEARCH is that it is
very slow on some server implementations (like Outlook365). This patch
avoid performing the SEARCH when possible by utilizing plain FETCH, and
only falls back to performing SEARCH when we must do UID FETCH.